daemon/pyproject.toml
[project]
name = "redbridge-daemon"
version = "0.0.1"
description = "Custom Stream Deck Plus daemon (no Elgato software)"
requires-python = ">=3.12"
dependencies = [
    "streamdeck>=0.9.5",
    "pillow>=10.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pywin32>=306 ; sys_platform == 'win32'",
    "pycaw>=20240210 ; sys_platform == 'win32'",
    "comtypes>=1.4 ; sys_platform == 'win32'",
]

[tool.uv]
package = false
